* adapt fisco 1.3 and 2.0
