+++ build.gradle
@@ -36,7 +36,7 @@ if (!gradle.startParameter.isOffline()) {
 
 group 'com.webank'
 
-version = "1.2.1"
+version = "1.3.0"
 
 // Specify JDK version - may vary in different scenarios
 sourceCompatibility = 1.8
@@ -92,16 +92,12 @@ List json = [
 ]
 
 List weidentity_contract = [
-        "com.webank:weidentity-contract-java:1.2.2"
+        "com.webank:weid-contract-java:1.2.3"
 ]
 
 List mysql_driver = [
         "mysql:mysql-connector-java:5.1.44",
         "org.apache.commons:commons-dbcp2:2.5.0"
-]
-
-List web3sdk = [
-        "org.fisco-bcos:web3sdk:1.2.5"
 ]
 
 List zxing = [
