+++ src/main/java/com/webank/weid/service/impl/AuthorityIssuerServiceImpl.java
@@ -44,10 +44,10 @@ import org.slf4j.LoggerFactory;
 import com.webank.weid.config.ContractConfig;
 import com.webank.weid.constant.ErrorCode;
 import com.webank.weid.constant.WeIdConstant;
-import com.webank.weid.contract.AuthorityIssuerController;
-import com.webank.weid.contract.AuthorityIssuerController.AuthorityIssuerRetLogEventResponse;
-import com.webank.weid.contract.SpecificIssuerController;
-import com.webank.weid.contract.SpecificIssuerController.SpecificIssuerRetLogEventResponse;
+import com.webank.weid.contract.v1.AuthorityIssuerController;
+import com.webank.weid.contract.v1.AuthorityIssuerController.AuthorityIssuerRetLogEventResponse;
+import com.webank.weid.contract.v1.SpecificIssuerController;
+import com.webank.weid.contract.v1.SpecificIssuerController.SpecificIssuerRetLogEventResponse;
 import com.webank.weid.protocol.base.AuthorityIssuer;
 import com.webank.weid.protocol.base.WeIdAuthentication;
 import com.webank.weid.protocol.request.RegisterAuthorityIssuerArgs;
