+++ src/main/java/com/webank/weid/service/impl/CptServiceImpl.java
@@ -46,8 +46,8 @@ import com.webank.weid.config.ContractConfig;
 import com.webank.weid.constant.ErrorCode;
 import com.webank.weid.constant.JsonSchemaConstant;
 import com.webank.weid.constant.WeIdConstant;
-import com.webank.weid.contract.CptController;
-import com.webank.weid.contract.CptController.UpdateCptRetLogEventResponse;
+import com.webank.weid.contract.v1.CptController;
+import com.webank.weid.contract.v1.CptController.UpdateCptRetLogEventResponse;
 import com.webank.weid.protocol.base.Cpt;
 import com.webank.weid.protocol.base.CptBaseInfo;
 import com.webank.weid.protocol.base.WeIdAuthentication;
