+++ src/main/java/com/webank/weid/service/impl/EvidenceServiceImpl.java
@@ -44,9 +44,9 @@ import org.slf4j.LoggerFactory;
 import com.webank.weid.config.ContractConfig;
 import com.webank.weid.constant.ErrorCode;
 import com.webank.weid.constant.WeIdConstant;
-import com.webank.weid.contract.Evidence;
-import com.webank.weid.contract.EvidenceFactory;
-import com.webank.weid.contract.EvidenceFactory.CreateEvidenceLogEventResponse;
+import com.webank.weid.contract.v1.Evidence;
+import com.webank.weid.contract.v1.EvidenceFactory;
+import com.webank.weid.contract.v1.EvidenceFactory.CreateEvidenceLogEventResponse;
 import com.webank.weid.protocol.base.Credential;
 import com.webank.weid.protocol.base.EvidenceInfo;
 import com.webank.weid.protocol.base.WeIdDocument;
