+++ src/main/java/com/webank/weid/util/DataToolUtils.java
@@ -44,6 +44,7 @@ import java.security.NoSuchAlgorithmException;
 import java.security.NoSuchProviderException;
 import java.security.SignatureException;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Hashtable;
 import java.util.Iterator;
@@ -992,6 +993,136 @@ public final class DataToolUtils {
     public static String bytes32ToString(Bytes32 bytes32) {
 
         return new String(bytes32.getValue(), StandardCharsets.UTF_8).trim();
+    }
+
+    /**
+     * string to byte.
+     *
+     * @param value stringData
+     * @return byte[]
+     */
+    public static byte[] stringToByteArray(String value) {
+        if (StringUtils.isBlank(value)) {
+            return new byte[1];
+        }
+        return value.getBytes(StandardCharsets.UTF_8);
+    }
+
+    /**
+     * string to byte32.
+     *
+     * @param value stringData
+     * @return byte[]
+     */
+    public static byte[] stringToByte32Array(String value) {
+        if (StringUtils.isBlank(value)) {
+            return new byte[32];
+        }
+
+        byte[] bytes = value.getBytes(StandardCharsets.UTF_8);
+        byte[] newBytes = new byte[32];
+
+        System.arraycopy(bytes, 0, newBytes, 0, bytes.length);
+        return newBytes;
+    }
+
+    /**
+     * string to byte32List.
+     *
+     * @param data stringData
+     * @param size size of byte32List
+     */
+    public static List<byte[]> stringToByte32ArrayList(String data, int size) {
+        List<byte[]> byteList = new ArrayList<>();
+
+        if (StringUtils.isBlank(data)) {
+            for (int i = 0; i < size; i++) {
+                byteList.add(new byte[32]);
+            }
+            return byteList;
+        }
+
+        byte[] dataBytes = data.getBytes(StandardCharsets.UTF_8);
+
+        if (dataBytes.length <= WeIdConstant.MAX_AUTHORITY_ISSUER_NAME_LENGTH) {
+            byte[] newBytes = new byte[32];
+            System.arraycopy(dataBytes, 0, newBytes, 0, dataBytes.length);
+            byteList.add(newBytes);
+        } else {
+            byteList = splitBytes(dataBytes, size);
+        }
+
+        if (byteList.size() < size) {
+            List<byte[]> addList = new ArrayList<>();
+            for (int i = 0; i < size - byteList.size(); i++) {
+                addList.add(new byte[32]);
+            }
+            byteList.addAll(addList);
+        }
+        return byteList;
+    }
+
+    private static synchronized List<byte[]> splitBytes(byte[] bytes, int size) {
+        List<byte[]> byteList = new ArrayList<>();
+        double splitLength =
+            Double.parseDouble(WeIdConstant.MAX_AUTHORITY_ISSUER_NAME_LENGTH + "");
+        int arrayLength = (int) Math.ceil(bytes.length / splitLength);
+        byte[] result = new byte[arrayLength];
+
+        int from = 0;
+        int to = 0;
+
+        for (int i = 0; i < arrayLength; i++) {
+            from = (int) (i * splitLength);
+            to = (int) (from + splitLength);
+
+            if (to > bytes.length) {
+                to = bytes.length;
+            }
+
+            result = Arrays.copyOfRange(bytes, from, to);
+            if (result.length < size) {
+                byte[] newBytes = new byte[32];
+                System.arraycopy(result, 0, newBytes, 0, result.length);
+                byteList.add(newBytes);
+            } else {
+                byteList.add(result);
+            }
+        }
+        return byteList;
+    }
+    /**
+     * convert bytesArrayList to Bytes32ArrayList.
+     * @param list byte size
+     * @param size size
+     * @return result
+     */
+    public static List<byte[]> bytesArrayListToBytes32ArrayList(List<byte[]> list, int size) {
+        List<byte[]> bytesList = new ArrayList<>();
+
+        if (list.isEmpty()) {
+            for (int i = 0; i < size; i++) {
+                bytesList.add(new byte[32]);
+            }
+            return bytesList;
+        }
+
+        for (byte[] bytes : list) {
+            if (bytes.length <= WeIdConstant.MAX_AUTHORITY_ISSUER_NAME_LENGTH) {
+                byte[] newBytes = new byte[32];
+                System.arraycopy(bytes, 0, newBytes, 0, bytes.length);
+                bytesList.add(newBytes);
+            }
+        }
+
+        if (bytesList.size() < size) {
+            List<byte[]> addList = new ArrayList<>();
+            for (int i = 0; i < size - bytesList.size(); i++) {
+                addList.add(new byte[32]);
+            }
+            bytesList.addAll(addList);
+        }
+        return bytesList;
     }
 
     /**
